refactor: 경로 탐색 기능 추상화
